[vnet_vxlan]: Refactor global vars and helper methods
