+++ tests/vxlan/vnet_config.py
@@ -1,0 +1,11 @@
+IPV6_VXLAN_TEST = False
+VXLAN_PORT = "13330"
+VXLAN_MAC = "00:aa:bb:cc:78:9a"
+APPLY_NEW_CONFIG = True
+CLEANUP = True
+DUT_VNET_SWITCH_CONFIG = "/tmp/vnet.switch.json"
+DUT_VNET_CONF = "/tmp/vnet.conf.json"
+DUT_VNET_ROUTE_CONFIG = "/tmp/vnet.route.json"
+DUT_VNET_INTF_CONFIG = "/tmp/vnet.intf.json"
+DUT_VNET_NBR_JSON = "/tmp/vnet.nbr.json"
+VNET_CONFIG = None
